Fixed Folder Structure and Added to index.js
