+++ package.json
@@ -2,7 +2,7 @@
   "name": "dubug-tool",
   "version": "1.0.1",
   "description": "Debug Tool",
-  "main": "src/debug.js",
+  "main": "./src/debug-tool.js",
   "directories": {
     "test": "test"
   },
@@ -30,6 +30,5 @@
     "chalk": "^1.1.3",
     "dotenv": "^4.0.0",
     "mocha": "^3.2.0",
-    "sequelize": "^3.30.2"
   }
 }
