+++ test/__test.js
@@ -1,0 +1,8 @@
+const expect = require('chai').expect;
+
+describe('Test Mocha', () => {
+
+  it('Our Tests should be run', () => {
+    expect(true).to.be.ok;
+  });
+});
